ml-model/requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
psycopg2-binary==2.9.10
joblib==1.4.2
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
